+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 - Rework parameters to avoid ARM GCC warnings for uninitialized var
 - Add support to define max-min option - with overflow
 - Add `.clang-format` draft
+- Add cron next option
+- Remove custom cron datetime structure